+++ src/lib/export.ts
@@ -6,11 +6,10 @@ import { stringify } from 'csv-stringify';
 import sqlString from 'sqlstring-sqlite';
 import Database from 'better-sqlite3';
 import mapSeries from 'promise-map-series';
-import untildify from 'untildify';
 
 import * as models from '../models/models.ts';
 import { openDb } from './db.ts';
-import { prepDirectory, generateFolderName } from './file-utils.ts';
+import { prepDirectory, generateFolderName, untildify } from './file-utils.ts';
 import { log, logWarning } from './log-utils.ts';
 import { formatCurrency, pluralize, setDefaultConfig } from './utils.ts';